chore: improve Editable exports
